convert usecase createChallenge
