+++ api/lib/domain/usecases/index.js
@@ -1,4 +1,3 @@
-export * from './create-challenge.js';
 export * from './create-competence.js';
 export * from './create-framework.js';
 export * from './create-mission.js';
